requirements.txt
discord.py>=2.3
aiosqlite
pydantic
apscheduler
pytest
